Tratamento de Dados/Descomplicando a Criação de Pacotes em Python/requirements.txt
numpy>=1.23.5
Pillow>=9.3.0